requirements.txt
# Core AI & API
groq>=0.9.0
openai>=1.0.0

# GUI Framework
PyQt5>=5.15.9
PyQtWebEngine>=5.15.6

# Data Visualization (embedded in PyQt5)
matplotlib>=3.7.1
seaborn>=0.12.2

# Data Processing
pandas>=2.0.0
numpy>=1.24.3
scipy>=1.10.1

# Data Schemas & Validation
pydantic>=2.0.0

# Environment & Config
python-dotenv>=1.0.0

# Excel & Reporting
openpyxl>=3.1.2
reportlab>=4.0.4
xlsxwriter>=3.1.2

# Utilities
colorama>=0.4.6
tqdm>=4.65.0
tabulate>=0.9.0
python-dateutil>=2.8.2
pytz>=2023.3
